Load Python type parsing lazily
