+++ tests/data/expected/main/cli_fast_paths/empty_formatters.txt
@@ -1,4 +1,6 @@
 {
   "generated": "# generated by datamodel-codegen:\n#   filename:  person.json\n\nfrom __future__ import annotations\nfrom pydantic import BaseModel, Field, conint\nfrom typing import Any\n\n\nclass Person(BaseModel):\n    firstName: str | None = Field(None, description=\"The person's first name.\")\n    lastName: str | None = Field(None, description=\"The person's last name.\")\n    age: conint(ge=0) | None = Field(None, description='Age in years which must be equal to or greater than zero.')\n    friends: list[Any] | None = None\n    comment: None = Field(None)",
-  "imported_format": false
+  "imported_format": false,
+  "imported_python_type_codec": false,
+  "imported_python_type_ir": false
 }
